LayoutTests/fast/canvas/webgl/array-setters-expected.txt

Test all permutations of WebGLArray setters to make sure values don't get truncated

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Regression test for https://bugs.webkit.org/show_bug.cgi?id=33350 : WebGLArray subclasses do the wrong conversion in indexSetter
Testing Int8Array
PASS webGLArray is array
PASS webGLArray[0] is array[0]
PASS webGLArray[1] is array[1]
PASS webGLArray[0] is 0
PASS webGLArray[1] is 0
PASS webGLArray[0] is array[0]
PASS webGLArray[1] is array[1]
PASS webGLArray[2] is 88
PASS webGLArray[3] is 99
PASS webGLArray[1] is 88
PASS webGLArray[2] is 99
PASS array[0] is 99
PASS array[1] is 99
PASS webGLArray[3] is 99
PASS webGLArray[4] is 77
PASS array[0] is 99
PASS array[1] is 99
Testing Uint8Array
PASS webGLArray is array
PASS webGLArray[0] is array[0]
PASS webGLArray[1] is array[1]
PASS webGLArray[0] is 0
PASS webGLArray[1] is 0
PASS webGLArray[0] is array[0]
PASS webGLArray[1] is array[1]
PASS webGLArray[2] is 88
PASS webGLArray[3] is 99
PASS webGLArray[1] is 88
PASS webGLArray[2] is 99
PASS array[0] is 99
PASS array[1] is 99
PASS webGLArray[3] is 99
PASS webGLArray[4] is 77
PASS array[0] is 99
PASS array[1] is 99
Testing Int16Array
PASS webGLArray is array
PASS webGLArray[0] is array[0]
PASS webGLArray[1] is array[1]
PASS webGLArray[0] is 0
PASS webGLArray[1] is 0
PASS webGLArray[0] is array[0]
PASS webGLArray[1] is array[1]
PASS webGLArray[2] is 88
PASS webGLArray[3] is 99
PASS webGLArray[1] is 88
PASS webGLArray[2] is 99
PASS array[0] is 99
PASS array[1] is 99
PASS webGLArray[3] is 99
PASS webGLArray[4] is 77
PASS array[0] is 99
PASS array[1] is 99
Testing Uint16Array
PASS webGLArray is array
PASS webGLArray[0] is array[0]
PASS webGLArray[1] is array[1]
PASS webGLArray[0] is 0
PASS webGLArray[1] is 0
PASS webGLArray[0] is array[0]
PASS webGLArray[1] is array[1]
PASS webGLArray[2] is 88
PASS webGLArray[3] is 99
PASS webGLArray[1] is 88
PASS webGLArray[2] is 99
PASS array[0] is 99
PASS array[1] is 99
PASS webGLArray[3] is 99
PASS webGLArray[4] is 77
PASS array[0] is 99
PASS array[1] is 99
Testing Int32Array
PASS webGLArray is array
PASS webGLArray[0] is array[0]
PASS webGLArray[1] is array[1]
PASS webGLArray[0] is 0
PASS webGLArray[1] is 0
PASS webGLArray[0] is array[0]
PASS webGLArray[1] is array[1]
PASS webGLArray[2] is 88
PASS webGLArray[3] is 99
PASS webGLArray[1] is 88
PASS webGLArray[2] is 99
PASS array[0] is 99
PASS array[1] is 99
PASS webGLArray[3] is 99
PASS webGLArray[4] is 77
PASS array[0] is 99
PASS array[1] is 99
Testing Uint32Array
PASS webGLArray is array
PASS webGLArray[0] is array[0]
PASS webGLArray[1] is array[1]
PASS webGLArray[0] is 0
PASS webGLArray[1] is 0
PASS webGLArray[0] is array[0]
PASS webGLArray[1] is array[1]
PASS webGLArray[2] is 88
PASS webGLArray[3] is 99
PASS webGLArray[1] is 88
PASS webGLArray[2] is 99
PASS array[0] is 99
PASS array[1] is 99
PASS webGLArray[3] is 99
PASS webGLArray[4] is 77
PASS array[0] is 99
PASS array[1] is 99
Testing Float32Array
PASS webGLArray is array
PASS webGLArray[0] is array[0]
PASS webGLArray[1] is array[1]
PASS webGLArray[0] is 0
PASS webGLArray[1] is 0
PASS webGLArray[0] is array[0]
PASS webGLArray[1] is array[1]
PASS webGLArray[2] is 88
PASS webGLArray[3] is 99
PASS webGLArray[1] is 88
PASS webGLArray[2] is 99
PASS array[0] is 99
PASS array[1] is 99
PASS webGLArray[3] is 99
PASS webGLArray[4] is 77
PASS array[0] is 99
PASS array[1] is 99
PASS successfullyParsed is true

TEST COMPLETE
